+++ kernel/c/timeranger2/src/timeranger2.h
@@ -424,7 +424,7 @@ PUBLIC int tranger2_write_user_flag(
     const char *key,        // In tranger2 ('key', '__t__', 'rowid') is required
     uint64_t __t__,
     uint64_t rowid,         // Must be real rowid in the file, not in topic global rowid
-    uint32_t user_flag
+    uint16_t user_flag
 );
 
 /*
@@ -437,7 +437,7 @@ PUBLIC int tranger2_set_user_flag(
     const char *key,        // In tranger2 ('key', '__t__', 'rowid') is required
     uint64_t __t__,
     uint64_t rowid,         // Must be real rowid in the file, not in topic global
-    uint32_t mask,
+    uint16_t mask,
     BOOL set
 );
